Extract common utility classes into a new Core package
