+++ .swiftpm/xcode/xcshareddata/xcschemes/Pillarbox-Package.xcscheme
@@ -253,6 +253,18 @@
             testExecutionOrdering = "random">
             <BuildableReference
                BuildableIdentifier = "primary"
+               BlueprintIdentifier = "CoreTests"
+               BuildableName = "CoreTests"
+               BlueprintName = "CoreTests"
+               ReferencedContainer = "container:">
+            </BuildableReference>
+         </TestableReference>
+         <TestableReference
+            skipped = "NO"
+            parallelizable = "YES"
+            testExecutionOrdering = "random">
+            <BuildableReference
+               BuildableIdentifier = "primary"
                BlueprintIdentifier = "CoreBusinessTests"
                BuildableName = "CoreBusinessTests"
                BlueprintName = "CoreBusinessTests"
