+++ Sources/Core/Core.docc/Core.md
@@ -1,0 +1,3 @@
+# ``Core``
+
+The Core library provides common tools which might be useful for applications using Pillarbox.
