+++ Sources/Player/DequeuePlayer.swift
@@ -9,7 +9,7 @@ import AVFoundation
 // TODO: Must later implement dequeue support for items to have playlist support in forward and backward directions.
 @MainActor
 public final class DequeuePlayer: AVQueuePlayer {
-    public override func seek(to time: CMTime, toleranceBefore: CMTime, toleranceAfter: CMTime, completionHandler: @escaping (Bool) -> Void) {
+    override public func seek(to time: CMTime, toleranceBefore: CMTime, toleranceAfter: CMTime, completionHandler: @escaping (Bool) -> Void) {
         NotificationCenter.default.post(name: .willSeek, object: self, userInfo: [
             SeekInfoKey.targetTime: time
         ])
