+++ Sources/Player/PlaybackProgress.swift
@@ -4,6 +4,7 @@
 //  License information is available from the LICENSE file.
 //
 
+import Core
 import Foundation
 
 /// Playback progress.
