+++ Tests/CoreTests/NotificationPublisherDeallocationTests.swift
@@ -4,7 +4,7 @@
 //  License information is available from the LICENSE file.
 //
 
-@testable import Player
+@testable import Core
 
 import Circumspect
 import Nimble
